Use higher precision in some places
Camera scale is now long double,
Winding uses fabsl, not really for the precision, but to
supress warning when complex uses long double

+++ includes/fractol.h
@@ -32,7 +32,7 @@ typedef struct s_img
 typedef struct s_camera
 {
 	int			iter;
-	double		scale;
+	long double	scale;
 	t_complex	center;
 	t_complex	top_left;
 	t_complex	step;
